fix(js): structuredClone 失败路径泄漏克隆对象；修正两处 CI 警告
用真实站点做回归测试时，www.qq.com、www.163.com、www.taobao.com 会在退出
时崩溃：

  quickjs.c:2704: JS_FreeRuntime: Assertion `list_empty(&rt->gc_obj_list)' failed

用 QuickJS 的泄漏转储（临时构建 -DENABLE_DUMPS，并在析构前设置
JS_DUMP_LEAKS）定位到一个引用计数为 1、无属性、无原型的裸对象——正好是
structuredClone 在克隆属性失败时遗弃的半成品克隆。

根因：CloneObject 在两条错误路径（读取属性抛异常、克隆属性抛
DataCloneError）直接 return 异常，没有释放本地持有的 `cloned`；虽然该对象
仍在 cx.seen 中（调用方会释放那一份 dup），但本地这一份永远留在
引用计数上，直到 runtime 销毁时触发断言。任何页面只要克隆一个含函数、
WeakMap 或抛错 getter 的对象（现代打包器很常见）就会命中。

修复：两条错误路径都先 JS_FreeValue(cloned) 再返回异常。

回归测试：PlatformApiTest.StructuredCloneReleasesThePartialCloneWhenCloningFails
覆盖函数值、抛错 getter、嵌套失败三种触发方式，并在其后验证正常克隆仍可用。
红检：临时回退修复后该测试即以同一断言失败（exit 134），恢复后通过。

顺带修复 CI（NEKO_WARNINGS_AS_ERRORS=ON）会拦下的两个警告：
- dns.cpp：ancount+nscount+arcount 先转 uint32_t 再相加（-Wsign-conversion）
- platform_binding.cpp：TextDecoderFinalizer 未使用的 rt 参数

验证：全量 ctest 1211/1211，-Werror 构建零警告，clang-format 227 文件干净；
GUI 截图通道 11 个常用站点全部 exit=0（含此前崩溃的三个）。

+++ src/network/src/dns.cpp
@@ -198,7 +198,12 @@ bool ParseDnsResponse(std::string_view message, DnsAnswer* out)
 
   // Answer records plus the authority/additional sections (skipped, but their
   // records must be walked so a malformed message is still rejected).
-  const std::uint32_t record_count = ancount + nscount + arcount;
+  // The three counts are summed in an unsigned type so the usual arithmetic
+  // conversions cannot produce a negative value (they are promoted to int
+  // otherwise, which -Wsign-conversion rejects).
+  const std::uint32_t record_count = static_cast<std::uint32_t>(ancount) +
+                                     static_cast<std::uint32_t>(nscount) +
+                                     static_cast<std::uint32_t>(arcount);
   bool first_ttl = true;
   for (std::uint32_t i = 0; i < record_count; ++i) {
     std::string name;
